+++ cercle-5/miniRT/srcs/primitives/objects/fdf/mesh.c
@@ -16,7 +16,7 @@ bool	fdf_init_mesh(t_mesh *mesh, size_t v_count, size_t i_count, const char *p)
 {
 	ft_memset(mesh, 0, sizeof(t_mesh));
 	mesh->vertices = ft_calloc(v_count, sizeof(t_vertex));
-	mesh->indices = ft_calloc(i_count, sizeof(unsigned int));
+	mesh->indices = ft_calloc(i_count, sizeof(size_t));
 	if (!mesh->vertices || !mesh->indices)
 	{
 		mesh_free(mesh);
